editoast: create connection pool refactor
