+++ editoast/src/main.rs
@@ -38,6 +38,7 @@ use client::{
 use editoast_schemas::infra::ElectricalProfileSetData;
 use editoast_schemas::rolling_stock::RollingStock;
 use editoast_schemas::train_schedule::TrainScheduleBase;
+use modelsv2::connection_pool::create_connection_pool;
 use modelsv2::{
     timetable::Timetable, timetable::TimetableWithTrains, train_schedule::TrainSchedule,
     train_schedule::TrainScheduleChangeset,
@@ -47,22 +48,16 @@ use opentelemetry_datadog::DatadogPropagator;
 use opentelemetry_sdk::propagation::TraceContextPropagator;
 use views::v2::train_schedule::{TrainScheduleForm, TrainScheduleResult};
 
-use crate::modelsv2::ConnectionConfig;
 use crate::modelsv2::Connection;
 use colored::*;
-use diesel::{sql_query, ConnectionError, ConnectionResult};
-use diesel_async::pooled_connection::deadpool::Pool;
-use diesel_async::pooled_connection::ManagerConfig;
+use diesel::sql_query;
 use diesel_async::RunQueryDsl;
 use diesel_json::Json as DieselJson;
 use editoast_schemas::infra::RailJson;
-use futures_util::future::BoxFuture;
-use futures_util::FutureExt;
 use infra_cache::InfraCache;
 use map::MapLayers;
 use modelsv2::electrical_profiles::ElectricalProfileSet;
 use modelsv2::prelude::*;
-use openssl::ssl::{SslConnector, SslMethod, SslVerifyMode};
 use opentelemetry::KeyValue;
 use opentelemetry_otlp::WithExportConfig as _;
 use opentelemetry_sdk::Resource;
@@ -76,7 +71,6 @@ use std::{env, fs};
 use thiserror::Error;
 use tracing::{error, info, warn};
 use tracing_subscriber::{layer::SubscriberExt as _, util::SubscriberInitExt as _, Layer as _};
-use url::Url;
 use validator::ValidationErrorsKind;
 use views::infra::InfraApiError;
 use views::search::{SearchConfig, SearchConfigFinder, SearchConfigStore};
@@ -184,7 +178,7 @@ async fn run() -> Result<(), Box<dyn Error + Send + Sync>> {
     init_tracing(EditoastMode::from_client(&client), &client.telemetry_config);
     let pg_config = client.postgres_config;
     let create_db_pool = || {
-        Ok::<_, Box<dyn Error + Send + Sync>>(Data::new(get_pool(
+        Ok::<_, Box<dyn Error + Send + Sync>>(Data::new(create_connection_pool(
             pg_config.url()?,
             pg_config.pool_size,
         )))
@@ -367,36 +361,6 @@ fn log_received_request(req: &ServiceRequest) {
         )
     };
     info!(target: "actix_logger", "{} RECEIVED", request_line);
-}
-
-fn establish_connection(config: &str) -> BoxFuture<ConnectionResult<Connection>> {
-    let fut = async {
-        let mut connector_builder = SslConnector::builder(SslMethod::tls()).unwrap();
-        connector_builder.set_verify(SslVerifyMode::NONE);
-        let tls = postgres_openssl::MakeTlsConnector::new(connector_builder.build());
-        let (client, conn) = tokio_postgres::connect(config, tls)
-            .await
-            .map_err(|e| ConnectionError::BadConnection(e.to_string()))?;
-        // The connection object performs the actual communication with the database,
-        // so spawn it off to run on its own.
-        tokio::spawn(async move {
-            if let Err(e) = conn.await {
-                error!("connection error: {}", e);
-            }
-        });
-        Connection::try_from(client).await
-    };
-    fut.boxed()
-}
-
-fn get_pool(url: Url, max_size: usize) -> ConnectionPool {
-    let mut manager_config = ManagerConfig::default();
-    manager_config.custom_setup = Box::new(establish_connection);
-    let manager = ConnectionConfig::new_with_config(url, manager_config);
-    Pool::builder(manager)
-        .max_size(max_size)
-        .build()
-        .expect("Failed to create pool.")
 }
 
 /// Create and run the server
